Move sealing related simulated inherent data provider to own mod
